Create Arduino PGA2311 Simple Communication
This is part of my project for making a remote controlled (ir remote based off my LG remote for my tv) to adjust volume (high quality). This is the basic serial or spi communcation using 4 pins from my controller to the chip pga 2311. Technically the code is ripped from a PGA 2310 but they have the same bla bla so it should work.